Add a program-global struct
This is a shamefully easy technique for keeping cross-cutting information. I
give both a mutable view and an immutable view of it. I try to discourage
mutable access by providing a really nice and shiny single-character identifier
for the immutable view: g. Who wouldn't want to change their logic for that?